Refactor: Implementado tratamento de exceção em toda a camada de Streaming serve, e Movie/Streaming Mapper foi implementado um metodo normalizarNome para padrozinar o nome salvo no banco de dados

+++ src/main/java/br/com/movieflix/service/StreamingService.java
@@ -2,8 +2,9 @@ package br.com.movieflix.service;
 
 import br.com.movieflix.controller.request.StreamingRequest;
 import br.com.movieflix.controller.response.StreamingResponse;
-import br.com.movieflix.entity.Category;
 import br.com.movieflix.entity.Streaming;
+import br.com.movieflix.exception.streamingException.StreamingAlreadyExistException;
+import br.com.movieflix.exception.streamingException.StremingNotFoundException;
 import br.com.movieflix.mapper.StreamingMapper;
 import br.com.movieflix.repository.StreamingRepository;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +12,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -32,35 +32,40 @@ public class StreamingService {
     public StreamingResponse findById(UUID id){
         return streamingRepository.findById(id)
                 .map(StreamingMapper::toStreamingResponse)
-                .orElse(null);
+                .orElseThrow(StremingNotFoundException::new);
     }
 
     @Transactional
     public StreamingResponse saveStreaming(StreamingRequest streamingRequest){
+        if (streamingRepository.existsStreamingByNameIgnoreCase(streamingRequest.name())){
+            throw new StreamingAlreadyExistException();
+        }
+
+
+
         Streaming streaming = streamingRepository.save(StreamingMapper.toStreaming(streamingRequest));
         return StreamingMapper.toStreamingResponse(streamingRepository.save(streaming));
     }
 
     @Transactional
     public void deleteById(UUID id) {
+        if (!streamingRepository.existsById(id)){
+            throw new StremingNotFoundException();
+        }
         streamingRepository.deleteById(id);
     }
 
     @Transactional(readOnly = true)
-    public Optional<Streaming> findByIdFromMovieList(UUID id){
-        return streamingRepository.findById(id);
+    public Streaming findByIdFromMovieList(UUID id){
+        return streamingRepository.findById(id).orElseThrow(StremingNotFoundException::new);
     }
 
     public StreamingResponse updateStreaming(UUID id, StreamingRequest streamingRequest) {
-        Optional<Streaming> streaming = streamingRepository.findById(id);
-
-        if (streaming.isPresent()){
-            if (streamingRequest.name() != null){
-                streaming.get().setName(streamingRequest.name());
-            }
-            return StreamingMapper.toStreamingResponse(streamingRepository.save(streaming.get()));
+        Streaming streaming = streamingRepository.findById(id).orElseThrow(StremingNotFoundException::new);
+        if (streamingRepository.existsStreamingByNameIgnoreCase(streamingRequest.name())){
+            throw new StreamingAlreadyExistException();
         }
-
-        return null;
+        streaming.setName(streamingRequest.name());
+        return StreamingMapper.toStreamingResponse(streamingRepository.save(streaming));
     }
 }
